move BaseTileLayer to geo-layers
