+++ modules/geo-layers/src/tile-layer/utils/tile-2d-header.js
