+++ test/modules/geo-layers/tile-layer/index.js
@@ -1,3 +1,4 @@
 import './tile-util.spec';
 import './viewport-util.spec';
 import './tile-layer.spec';
+import './tileset-2d.spec';
